Add a TWidget object
