+++ src/phink/web/ui/widget/plugin/plugin_renderer.php
@@ -16,14 +16,16 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
  
- namespace Phink\Web\UI\Widget\Plugin;
+namespace Phink\Web\UI\Widget\Plugin;
+
+use Phink\Web\UI\Widget\TWidget;
 
 /**
  * Description of tgrid
  *
  * @author david
  */
-abstract class TPluginRenderer extends \Phink\MVC\TPartialController
+abstract class TPluginRenderer extends TWidget
 {
     use \Phink\Web\UI\THtmlPattern;
     use \Phink\Data\UI\TDataBinder;
