+++ src/widgets/pager/controllers/pager.class.php
@@ -20,32 +20,15 @@ namespace Phink\Widgets\Pager;
 
 use Phink\TAutoloader;
 use Phink\Registry\TRegistry;
+use Phink\Web\UI\Widget\TWidget;
 
-class TPager extends \Phink\MVC\TPartialController
+class TPager extends TWidget
 {
 
     protected $pageCount;
-    protected $caption;
     protected $currentPage;
     protected $pageNum;
-    protected $statement;
-    protected $onclick;
-    protected $script;
     protected $pagerJS;
-    protected $forThis;
-    protected $forView;
-    protected $forCtrl;
-    protected $forApp;
-
-    public function setStatement($value) : void
-    {
-        $this->statement = $value;
-    }
-
-    public function setCaption($value) : void
-    {
-        $this->caption = $value;
-    }
 
     public function setPageCount($value) : void
     {
@@ -62,21 +45,6 @@ class TPager extends \Phink\MVC\TPartialController
         $this->pageNum = $value;
     }
     
-    public function setOnclick($value) : void
-    {
-        $this->onclick = $value;
-    }
-    
-    public function setFor($value) : void
-    {
-        $this->forThis = $value;
-    }
-
-    public function getCacheFilename() : string
-    {
-        return SRC_ROOT . REL_RUNTIME_DIR . str_replace(DIRECTORY_SEPARATOR, '_', $this->path . $this->forThis . 'pager' . CLASS_EXTENSION);
-    }
-
     public function init() : void
     {
         $forControl = $this->parent->getChildById($this->forThis);
